Fix import of NVM bundle file or scene folder with .cam files on Windows
Importing an NVM bundle file or a scene folder with .cam files where the path
is absolute doesn't work on Windows because the tokenization in
tex::generate_texture_views() breaks the absolute path which contains a ":"
on Windows (i.e. C:\foo) resulting in tok[0] only containing a drive letter
and not the full absolute path on Windows. This is fixed by simply not
tokenizing the given path which automatically makes every option work
correctly on both Windows and Unix-like OSs. Really fixes nmoehrle/mvs-texturing#37.
Thanks to @hushanming for reporting the bug!