openmp: Switch from `critical` to `ordered` for determinism.
When adding to vectors (e.g. using `push_back()`), `ordered` in contrast to
`critical` ensures that ordering of operations is the same as if run
single-threadedly, and it ensures that the result is the same on every run
of the program.

This allows to generate deterministic results: Run with same inputs,
byte-identical outputs are to be produced, independent of threading.

I have checked using `time` on a 6-core machine that the changes have no
significant impact on performance; this is expected because the critical regions
are very small (usually adding small pointers to vectors).

One location remains that still uses `critical` because `omp ordered` cannot be
used inside `pragma omp parallel`, only inside `pragma omp parallel for`;
this is likely because of the thread-local variable `projected_face_view_infos`
being intended as a per-thread intermediate buffer; more effort needs to be put
into how that can be put back into order.
I've added a TODO for this.
I haven't yet observed nondeterminism due to this, but may as well have been
lucky.